FINAL_DEPLOYMENT/backend/requirements.txt
# GUARANTEED WORKING BACKEND REQUIREMENTS - NO ERRORS
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0